# pkg/webapi/docs/gen/swagger.yaml
basePath: /api
definitions:
...
  dto.MapRouteRequest:
    properties:
      color:
        type: string
      geojson:
        type: object
      name:
        type: string
    type: object
  dto.MapRouteResponse:
    properties:
      color:
        type: string
      geojson:
        type: object
      id:
        type: integer
      name:
        type: string
      point_count:
        type: integer
    type: object
...
  webtypes.ErrorResponse:
    properties:
      error:
        type: string
    type: object
info:
  contact: {}
  description: REST API for graywolf configuration and control.
  title: Graywolf Management API
  version: "1.0"
...
paths:
...
  /map-routes:
    get:
      operationId: listMapRoutes
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.MapRouteResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/webtypes.ErrorResponse'
      security:
        - CookieAuth: []
      summary: List map routes
      tags:
        - map-routes
    post:
      consumes:
        - application/json
      operationId: createMapRoute
      parameters:
        - description: Route definition
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/dto.MapRouteRequest'
      produces:
        - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.MapRouteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/webtypes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/webtypes.ErrorResponse'
      security:
        - CookieAuth: []
      summary: Create map route
      tags:
        - map-routes
  /map-routes/{id}:
    delete:
      operationId: deleteMapRoute
      parameters:
        - description: Route id
          in: path
          name: id
          required: true
          type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/webtypes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/webtypes.ErrorResponse'
      security:
        - CookieAuth: []
      summary: Delete map route
      tags:
        - map-routes
    get:
      operationId: getMapRoute
      parameters:
        - description: Route id
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MapRouteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/webtypes.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/webtypes.ErrorResponse'
      security:
        - CookieAuth: []
      summary: Get map route
      tags:
        - map-routes
    put:
      consumes:
        - application/json
      operationId: updateMapRoute
      parameters:
        - description: Route id
          in: path
          name: id
          required: true
          type: integer
        - description: Route definition
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/dto.MapRouteRequest'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MapRouteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/webtypes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/webtypes.ErrorResponse'
      security:
        - CookieAuth: []
      summary: Update map route
      tags:
        - map-routes
...
schemes:
  - http
securityDefinitions:
  CookieAuth:
    description: |-
      Session cookie issued by POST /api/auth/login.
      Swagger 2.0 lacks a native `in: cookie` apiKey
      location, so the spec models the same credential
      as a `Cookie:` request header. Browsers send this
      automatically once the session cookie is set; the
      session cookie is named `graywolf_session`.
    in: header
    name: Cookie
    type: apiKey
swagger: "2.0"